+++ src/ReadFrom.php
@@ -4,9 +4,10 @@ declare(strict_types=1);
 
 namespace Thesis\ByteOrder;
 
+use BcMath\Number;
 use Thesis\ByteReader\Reader;
 use Thesis\ByteReader\UnexpectedEof;
-use Thesis\Endian\endian;
+use Thesis\Endian\Order;
 
 /**
  * @api
@@ -16,54 +17,53 @@ interface ReadFrom extends Reader
     /**
      * @throws UnexpectedEof
      */
-    public function readInt8(endian $endian = endian::network): int;
+    public function readInt8(Order $order = Order::Network): int;
 
     /**
      * @return non-negative-int
      * @throws UnexpectedEof
      */
-    public function readUint8(endian $endian = endian::network): int;
+    public function readUint8(Order $order = Order::Network): int;
 
     /**
      * @throws UnexpectedEof
      */
-    public function readInt16(endian $endian = endian::network): int;
-
-    /**
-     * @return non-negative-int
-     * @throws UnexpectedEof
-     */
-    public function readUint16(endian $endian = endian::network): int;
-
-    /**
-     * @throws UnexpectedEof
-     */
-    public function readInt32(endian $endian = endian::network): int;
+    public function readInt16(Order $order = Order::Network): int;
 
     /**
      * @return non-negative-int
      * @throws UnexpectedEof
      */
-    public function readUint32(endian $endian = endian::network): int;
+    public function readUint16(Order $order = Order::Network): int;
 
     /**
      * @throws UnexpectedEof
      */
-    public function readInt64(endian $endian = endian::network): int;
+    public function readInt32(Order $order = Order::Network): int;
 
     /**
      * @return non-negative-int
      * @throws UnexpectedEof
      */
-    public function readUint64(endian $endian = endian::network): int;
+    public function readUint32(Order $order = Order::Network): int;
 
     /**
      * @throws UnexpectedEof
      */
-    public function readFloat(endian $endian = endian::network): float;
+    public function readInt64(Order $order = Order::Network): Number;
 
     /**
      * @throws UnexpectedEof
      */
-    public function readDouble(endian $endian = endian::network): float;
+    public function readUint64(Order $order = Order::Network): Number;
+
+    /**
+     * @throws UnexpectedEof
+     */
+    public function readFloat(Order $order = Order::Network): float;
+
+    /**
+     * @throws UnexpectedEof
+     */
+    public function readDouble(Order $order = Order::Network): float;
 }